+++ source/frontend/src/features/admin/financial-codes/list/FinancialCodeListView.tsx
@@ -1,6 +1,7 @@
 import { Button } from 'components/common/buttons/Button';
 import { TableSort } from 'components/Table/TableSort';
 import { Roles } from 'constants/roles';
+import { useFinancialCodeRepository } from 'hooks/repositories/useFinancialCodeRepository';
 import useKeycloakWrapper from 'hooks/useKeycloakWrapper';
 import orderBy from 'lodash/orderBy';
 import { Api_FinancialCode } from 'models/api/FinancialCode';
@@ -11,7 +12,6 @@ import { FaPlus } from 'react-icons/fa';
 import { useHistory } from 'react-router';
 import styled from 'styled-components';
 
-import { useFinancialCodeRepository } from '../hooks/useFinancialCodeRepository';
 import {
   defaultFinancialCodeFilter,
   FinancialCodeFilter,
