+++ source/frontend/src/features/admin/financial-codes/update/UpdateFinancialCodeContainer.tsx
@@ -2,6 +2,7 @@ import { AxiosError } from 'axios';
 import { H1 } from 'components/common/styles';
 import LoadingBackdrop from 'components/maps/leaflet/LoadingBackdrop/LoadingBackdrop';
 import { FinancialCodeTypes } from 'constants/index';
+import { useFinancialCodeRepository } from 'hooks/repositories/useFinancialCodeRepository';
 import { IApiError } from 'interfaces/IApiError';
 import { Api_FinancialCode } from 'models/api/FinancialCode';
 import React, { useEffect, useState } from 'react';
@@ -10,7 +11,6 @@ import { useHistory } from 'react-router-dom';
 import { toast } from 'react-toastify';
 import styled from 'styled-components';
 
-import { useFinancialCodeRepository } from '../hooks/useFinancialCodeRepository';
 import { UpdateFinancialCodeYupSchema } from './UpdateFinancialCodeYupSchema';
 
 export interface IUpdateFinancialCodeFormProps {
